+++ src/shift_left/src/shift_left/core/utils/ksql_code_agent.py
@@ -21,9 +21,13 @@ class KsqlFlinkSql(BaseModel):
 class FlinkSql(BaseModel):
     ddl_sql_input: str
     dml_sql_input: str
-    error_message: str
     flink_ddl_output: str
     flink_dml_output: str
+
+class FlinkSqlForRefinement(BaseModel):
+    sql_input: str
+    error_message: str
+    flink_output: str
 
 
 class KsqlToFlinkSqlAgent:
@@ -42,19 +46,24 @@ class KsqlToFlinkSqlAgent:
         compute_pool_id = config.get('flink').get('compute_pool_id')
         statement = None
         if sql_to_validate:
-            statement_name = "syntax-check-flink-ddl"
+            statement_name = "syntax-check"
             delete_statement_if_exists(statement_name)
             statement = post_flink_statement(compute_pool_id, statement_name, sql_to_validate)
             print(f"CC Flink Statement: {statement}")
             logger.info(f"CC Flink Statement: {statement}")
-        
-        if statement and isinstance(statement, Statement) and statement.status:
-            if statement.status.phase == "RUNNING":
-                return True, ""
+            if statement and statement.status:
+                if statement.status.phase == "RUNNING":
+                    # Stop the statement as not needed anymore
+                    delete_statement_if_exists(statement_name)
+                    return True, ""
+                elif statement.status.phase == "COMPLETED":
+                    return True, statement.status.detail
+                else:
+                    return False, statement.status.detail
             else:
-                return False, statement.status.detail
+                return False, "No statement found"
         else:
-            return False, statement
+            return False, "No sql to validate"
         
     def _load_prompts(self):
         fname = importlib.resources.files("shift_left.core.utils.prompts.ksql_fsql").joinpath("translator.txt")
@@ -79,7 +88,7 @@ class KsqlToFlinkSqlAgent:
                         options={'temperature': 0}
                         )
         obj_response = KsqlFlinkSql.model_validate_json(response.message.content)
-        return obj_response.flink_dml_output, obj_response.flink_ddl_output
+        return obj_response.flink_ddl_output,  obj_response.flink_dml_output
 
     def _mandatory_validation_agent(self, ddl_sql: str, dml_sql: str) -> Tuple[str, str]:
         """
@@ -96,7 +105,7 @@ class KsqlToFlinkSqlAgent:
                         options={'temperature': 0}
                         )
         obj_response = FlinkSql.model_validate_json(response.message.content)
-        return obj_response.flink_dml_output, obj_response.flink_ddl_output
+        return obj_response.flink_ddl_output, obj_response.flink_dml_output
 
     def _refinement_agent(self, sql: str, 
                               history: str,
@@ -108,12 +117,12 @@ class KsqlToFlinkSqlAgent:
             {"role": "user", "content": refinement_prompt_template.format(sql_input=sql, history=history, error_message=error_message)}
         ]
         response= chat(model=self.model_name, 
-                        format=FlinkSql.model_json_schema(), 
+                        format=FlinkSqlForRefinement.model_json_schema(), 
                         messages=messages, 
                         options={'temperature': 0}
                         )
-        obj_response = FlinkSql.model_validate_json(response.message.content)
-        return obj_response.flink_dml_output
+        obj_response = FlinkSqlForRefinement.model_validate_json(response.message.content)
+        return obj_response.flink_output
 
 
     def _process_semantic_validation(self, sql: str) -> str:
@@ -124,18 +133,20 @@ class KsqlToFlinkSqlAgent:
 
     def translate_from_ksql_to_flink_sql(self, ksql: str, validate: bool = False) -> Tuple[str, str]:
         """
-        Entry point to translate ksql to flink sql using LLM Agents
+        Entry point to translate ksql to flink sql using LLM Agents. This is the workflow implementation.
         """
         
-        dml_sql, ddl_sql = self._translator_agent(ksql)
-        print(f"Done with translator agent, the flink DDL sql is:\n {ddl_sql}\nand DML: {dml_sql}")
-        dml_sql, ddl_sql = self._mandatory_validation_agent(ddl_sql, dml_sql)
-        print(f"Done with mandatory validation agent updated flink DDL sql is:\n {ddl_sql}\nand DML: {dml_sql}")
+        ddl_sql, dml_sql = self._translator_agent(ksql)
+        print(f"Done with translator agent, the flink DDL sql is:\n {ddl_sql}\nand DML: {dml_sql if dml_sql else 'empty'}\n2/ Start mandatory validation agent...")
+        logger.info(f"Done with translator agent, the flink DDL sql is:\n {ddl_sql}\nand DML: {dml_sql if dml_sql else 'empty'}")
+        ddl_sql, dml_sql = self._mandatory_validation_agent(ddl_sql, dml_sql)
+        print(f"Done with mandatory validation agent updated flink DDL sql is:\n {ddl_sql}\nand DML: {dml_sql if dml_sql else 'empty'}")
+        logger.info(f"Done with mandatory validation agent, updated flink DDL sql is:\n {ddl_sql}\nand DML: {dml_sql if dml_sql else 'empty'}")
         if validate:
-            print("Start validating the flink SQLs using Confluent Cloud for Flink, do you want to continue? (y/n)")
+            print("3/ Start validating the flink SQLs using Confluent Cloud for Flink, do you want to continue? (y/n)")
             answer = input()
             if answer != "y":
-                return dml_sql, ddl_sql
+                return ddl_sql, dml_sql
             ddl_sql, validated = _iterate_on_validation(self, ddl_sql)
             if validated:
                 final_ddl = self._process_semantic_validation(ddl_sql)
@@ -145,15 +156,13 @@ class KsqlToFlinkSqlAgent:
                     return final_ddl, final_dml
                 else:
                     return final_ddl, dml_sql   
-            else:
-                return ddl_sql, dml_sql
         return ddl_sql, dml_sql
 
 
 def _iterate_on_validation(self, translated_sql: str) -> Tuple[str, bool]:
     sql_validated = False
     iteration_count = 0
-    agent_history = []
+    agent_history = [{"agent": "refinement", "sql": translated_sql}]
     while not sql_validated and iteration_count < 3:
         iteration_count += 1
         sql_validated, status_detail = self._validate_flink_sql_on_cc(translated_sql)
@@ -163,10 +172,10 @@ def _iterate_on_validation(self, translated_sql: str) -> Tuple[str, bool]:
             translated_sql = self._refinement_agent(translated_sql, str(agent_history), status_detail)
             print(f"Refined sql:\n {translated_sql}")
             agent_history.append({"agent": "refinement", "sql": translated_sql})
-        print("do you want to continue? (y/n) or you continue with the generated sql?")
-        answer = input()
-        if answer != "y":
-            return translated_sql, sql_validated
+            print("do you want to continue? (y/n) or you continue with the generated sql?")
+            answer = input()
+            if answer != "y":
+                return translated_sql, sql_validated
     return translated_sql, sql_validated