+++ src/shift_left/tests/data/ksql-project/flink-references/basic_table_stream/sql-scripts/ddl.basic_table_stream.sql
@@ -1,17 +1,19 @@
-CREATE TABLE IF NOT EXISTS basic_table (
-  kpiId VARCHAR,
-  kpiName VARCHAR,
-  kpiStatus VARCHAR,
+CREATE TABLE IF NOT EXISTS BASIC_TABLE_STREAM (
+  kpiName STRING,
+  kpiStatus STRING,
   highMetricThreshold INT,
   lowMetricThreshold INT,
-  PRIMARY KEY(kpiId) NOT ENFORCED
-) DISTRIBUTED BY HASH(kpiId) INTO 6 BUCKETS
+  networkService STRING,
+  elementType STRING,
+  interfaceName STRING,
+  dbTable STRING,
+  PRIMARY KEY(kpiName) NOT ENFORCED
+) DISTRIBUTED BY HASH(kpiName) INTO 1 BUCKETS
 WITH (
-  'changelog.mode' = 'upsert',
+  'changelog.mode' = 'append',
   'key.avro-registry.schema-context' = '.flink-dev',
   'value.avro-registry.schema-context' = '.flink-dev',
-  'key.format' = 'avro-registry',
-  'value.format' = 'avro-registry',
+  'value.format' = 'json-registry',
   'kafka.retention.time' = '0',
   'kafka.producer.compression.type' = 'snappy',
   'scan.bounded.mode' = 'unbounded',