+++ src/shift_left/tests/data/ksql-project/staging/ut/map_location/sql-scripts/dml.map_location.sql
@@ -1,5 +1,0 @@
-INSERT INTO map_location
-SELECT 
--- part to select stuff
-FROM src_table
-WHERE -- where condition or remove it